+++ submissions/CodeMapper/src/FreeMindLabs.SemanticKernel.Plugins.CodeMapper/CategoryItem.cs
@@ -1,0 +1,9 @@
+﻿// Semantic Kernel Hackathon 2 - Code Mapper by Free Mind Labs.
+
+namespace FreeMindLabs.SemanticKernel.Plugins.CodeMapper;
+
+public class CategoryItem
+{
+    public string Id { get; set; }
+    public string Description { get; set; }
+}
