+++ submissions/CodeMapper/src/FreeMindLabs.SemanticKernel.Plugins.CodeMapper/CodeItem.cs
@@ -1,0 +1,12 @@
+﻿// Semantic Kernel Hackathon 2 - Code Mapper by Free Mind Labs.
+
+namespace FreeMindLabs.SemanticKernel.Plugins.CodeMapper;
+
+public class CodeItem
+{
+    public string Id { get; set; }
+    public string CategoryId { get; set; }
+    public string CategoryDescription { get; set; }
+    public string Code { get; set; }
+    public string Description { get; set; }
+}
